+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.4.3.9000
 
+* New `src_memdb()` which is a session-local in-memory SQLite db.
+
 * `distinct()` now only keeps the distinct variables. If you want to return
   all variables (using the first row for non-distinct values) use
   `.keep_all = TRUE` (#1110).